:building_construction: 컴포넌트별로 코드를 분리하고 기능에 따라 코드 재분류
